+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -3599,7 +3599,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01KKBPGBNM5K9KET9KXVXJFXFA"),
+  id: z.string().default("01KKBQR21VEVZP4DFXXE23DZ7A"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),