+++ packages/orq-rc/RELEASES.md
@@ -7579,3 +7579,13 @@ Based on:
 - [typescript v4.5.0-rc.45] packages/orq-rc
 ### Releases
 - [NPM v4.5.0-rc.45] https://www.npmjs.com/package/@orq-ai/node/v/4.5.0-rc.45 - packages/orq-rc
+
+## 2026-03-10 11:25:55
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.750.0 (2.856.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.5.0-rc.46] packages/orq-rc
+### Releases
+- [NPM v4.5.0-rc.46] https://www.npmjs.com/package/@orq-ai/node/v/4.5.0-rc.46 - packages/orq-rc